Decouple the FJS MCP server from CAS
Move the generic, transport-agnostic MCP protocol implementation from
fjs/mcp/ to fjs/protocol/mcp/, and the JSON-RPC 2.0 envelope/dispatcher
it layers on from fjs/media/json/rpc/ to fjs/protocol/json_rpc/. The
vacated fjs/mcp/ now holds the FJS MCP server itself (formerly
fjs/cas/mcp/), split into fjs/mcp/module.f.ts (the composition root:
config, CLI entry point, registry wiring) and fjs/mcp/cas/ (the cas_*
tool registry), so fjs/mcp/ matches the `fjs mcp` CLI command 1:1 and
isn't inherently CAS-specific. The Evo tool registry moves from
fjs/cas/evo/mcp/ to fjs/mcp/evo/.
