+++ fjs/cas/README.md
@@ -6,7 +6,7 @@ raw bytes only, sharded by cBase32 hash under `~/.cas/` — and never records a
 mutable pointer or a per-blob type tag; both are recovered at the edges that
 need them.
 
-- [fjs/cas/mcp](mcp/) — an MCP front end for agents, including read-time
+- [fjs/mcp](../mcp/) — an MCP front end for agents, including read-time
   media-type detection: `cas_add` / `cas_get` / `cas_list` plus
   [fjs/cas/evo](evo/)'s `evo_list` / `evo_head` / `evo_revision` / `evo_add`,
   all served from one process.
