+++ fjs/cas/evo/todo/evo-list-archived-filtering.md
@@ -67,4 +67,4 @@ the default result set and is therefore a breaking change.
 
 - [PR #1386](https://github.com/functionalscript/functionalscript/pull/1386)
 - [fjs/cas/evo/README.md](../README.md) — Evo cache and current-head behavior.
-- [fjs/cas/evo/mcp/](../mcp/) — MCP adapter for Evo commands.
+- [fjs/mcp/evo/](../../../mcp/evo/) — MCP adapter for Evo commands.
