+++ fjs/cas/todo/66n-cas-get-response-shape.md
@@ -12,7 +12,7 @@
 
 ### Problem
 
-The `cas_get` handler in `fjs/cas/mcp/module.f.ts` builds a `Meta` record from
+The `cas_get` handler in `fjs/mcp/cas/module.f.ts` builds a `Meta` record from
 a `fjs/media/type` detector verdict (`{ length, mime_type, type }`) plus `uri`
 in two places:
 
@@ -60,7 +60,7 @@ only the two genuinely different tail actions remain at the call sites.
       `cas_get` handler and route both `Meta` constructions through it.
 - [ ] Evaluate folding the two `collectRead` + `detectDialect` arms into a
       shared refine step; apply if it stays readable.
-- [ ] `npx tsc`, `fjs t`; `fjs/cas/mcp/proof.f.ts` passes with full coverage
+- [ ] `npx tsc`, `fjs t`; `fjs/mcp/proof.f.ts` passes with full coverage
       and byte-identical responses.
 
 ### Related
