+++ fjs/mcp/evo/proof.f.ts
@@ -1,13 +1,13 @@
-import { assert, assertEq } from '../../../asserts/module.f.ts'
-import { fileCas } from '../../module.f.ts'
-import { sha256 } from '../../../crypto/sha2/module.f.ts'
-import { emptyState, virtual } from '../../../effects/node/virtual/module.f.ts'
-import { vec8 } from '../../../types/bit_vec/module.f.ts'
-import { vecToCBase32 } from '../../../basen/cbase32/module.f.ts'
-import { initEvo, evo, type Evo } from '../module.f.ts'
+import { assert, assertEq } from '../../asserts/module.f.ts'
+import { fileCas } from '../../cas/module.f.ts'
+import { sha256 } from '../../crypto/sha2/module.f.ts'
+import { emptyState, virtual } from '../../effects/node/virtual/module.f.ts'
+import { vec8 } from '../../types/bit_vec/module.f.ts'
+import { vecToCBase32 } from '../../basen/cbase32/module.f.ts'
+import { initEvo, evo, type Evo } from '../../cas/evo/module.f.ts'
 import { evoToolRegistry } from './module.f.ts'
-import type { ToolEntry, ToolsCallResult } from '../../../mcp/module.f.ts'
-import type { Operation } from '../../../effects/module.f.ts'
+import type { ToolEntry, ToolsCallResult } from '../../protocol/mcp/module.f.ts'
+import type { Operation } from '../../effects/module.f.ts'
 
 const home = '/home/user'
