+++ fjs/media/json/todo/stringify-sorted-canonical.md
@@ -11,7 +11,7 @@ re-named at every use site instead of existing once under one name.
 
 Source modules:
 
-- `fjs/mcp/stdio/module.f.ts:49` — `const stringifyJson = stringify(sort)`
+- `fjs/protocol/mcp/stdio/module.f.ts:49` — `const stringifyJson = stringify(sort)`
 - `fjs/djs/module.f.ts:44` — `stringify(sort)(result[1])` inline
 
 Proof files (each binds its own alias: `jsonStr`, `str`, `stringify`,
@@ -24,7 +24,7 @@ Proof files (each binds its own alias: `jsonStr`, `str`, `stringify`,
   `fjs/types/sorted_set/proof.f.ts:10`, `fjs/types/list/proof.f.ts:37`
 - `fjs/text/ascii/proof.f.ts:6`, `fjs/text/utf8/proof.f.ts:8`,
   `fjs/text/utf16/proof.f.ts:15`
-- `fjs/media/json/parser/proof.f.ts:13`, `fjs/mcp/stdio/proof.f.ts:13`
+- `fjs/media/json/parser/proof.f.ts:13`, `fjs/protocol/mcp/stdio/proof.f.ts:13`
 - `fjs/bnf/data/proof.f.ts` (10 inline calls), `fjs/djs/parser/proof.f.ts:306`,
   `fjs/djs/serializer/proof.f.ts:47`
 
@@ -57,7 +57,7 @@ applications once.
 - [ ] Add `stringifySorted` to `fjs/media/json/module.f.ts` with proof
       coverage in `fjs/media/json/proof.f.ts` (which itself calls
       `stringify(sort)` seven times today).
-- [ ] Migrate the two source-module sites (`fjs/mcp/stdio/module.f.ts`,
+- [ ] Migrate the two source-module sites (`fjs/protocol/mcp/stdio/module.f.ts`,
       `fjs/djs/module.f.ts`), then the proof files.
 - [ ] Run `npx tsc` and `fjs t`.
