+++ fjs/protocol/json_rpc/module.f.ts
@@ -14,11 +14,11 @@
  *
  * @module
  */
-import { number, string, or, option } from '../../../types/rtti/module.f.ts'
-import type { Ts } from '../../../types/rtti/ts/module.f.ts'
-import { validate } from '../../../types/rtti/validate/module.f.ts'
-import type { Result } from '../../../types/result/module.f.ts'
-import { unknown, type Unknown } from '../module.f.ts'
+import { number, string, or, option } from '../../types/rtti/module.f.ts'
+import type { Ts } from '../../types/rtti/ts/module.f.ts'
+import { validate } from '../../types/rtti/validate/module.f.ts'
+import type { Result } from '../../types/result/module.f.ts'
+import { unknown, type Unknown } from '../../media/json/module.f.ts'
 
 export const jsonrpc = '2.0' as const
