+++ fjs/protocol/json_rpc/proof.f.ts
@@ -1,6 +1,6 @@
-import { assert, assertEq } from '../../../asserts/module.f.ts'
-import { ok, error } from '../../../types/result/module.f.ts'
-import { validate } from '../../../types/rtti/validate/module.f.ts'
+import { assert, assertEq } from '../../asserts/module.f.ts'
+import { ok, error } from '../../types/result/module.f.ts'
+import { validate } from '../../types/rtti/validate/module.f.ts'
 import {
     error as errorSchema,
     decodeRequest,
