+++ fjs/protocol/mcp/todo/output-type.md
